shut up the following warning on MacOS X Lion, pointed out by Soda-san (n-soda@) {{{ gfiops.c:526: warning: format '%06ld' expects type 'long int', but argument 7 has type '__darwin_suseconds_t' gfiops.c:526: warning: format '%06ld' expects type 'long int', but argument 7 has type '__darwin_suseconds_t' }}}
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/trunk@5844 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ bench/gfiops/gfiops.c
@@ -523,7 +523,7 @@ result(const char *name, struct timeval *t1, struct timeval *t2,
 	printf("%9s: %6d /sec: %d proc * %4d: "
 	       "%2ld.%06ld sec: %9.3f usec * %d\n",
 	       name, num_persec, nprocs, ntimes,
-	       (long int) t2->tv_sec, t2->tv_usec, usec, n);
+	       (long int) t2->tv_sec, (long int) t2->tv_usec, usec, n);
 }
 
 static int
